requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
decorator==4.4.2
future==0.18.2
graphviz==0.14.2
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.2.0
livereload==2.6.3
lunr==0.5.8
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mdutils==1.3.0
mkdocs==1.1.2
mkdocs-windmill-dark==0.2.0
mypy-extensions==0.4.3
networkx==2.5
nltk==3.5
numpy==1.19.2
pathspec==0.8.0
Pillow==8.0.1
pygraphviz==1.6
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.10.23
six==1.15.0
toml==0.10.1
tornado==6.0.4
tqdm==4.51.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
